Refactor attestation to a separate project
